change next for common react
