--- Client/.babelrc
@@ -1,18 +1,0 @@
-{
-    "presets": ["next/babel"],
-    "plugins": [
-        [
-            "module-resolver",
-            {
-                "root": ["./"],
-                "alias": {
-                    "@config": "./config",
-                    "@utils": "./src/utils",
-                    "@redux": "./src/redux",
-                    "@components": "./src/components",
-                    "@public": "./public"
-                }
-            }
-        ]
-    ]
-}
