--- Client/next.config.js
@@ -1,9 +1,0 @@
-// Dependencies
-const withSass = require('@zeit/next-sass');
-
-module.exports = withSass({
-    cssModules: true,
-    cssLoaderOptions: {
-        localIdentName: '[local]'
-    }
-});
