--- Client/package.json
@@ -1,40 +1,0 @@
-{
-    "name": "group_chat_client",
-    "version": "1.0.0",
-    "description": "",
-    "main": "index.js",
-    "scripts": {
-        "dev": "babel-node src/server.js",
-        "build": "next build",
-        "start": "cross-env NODE_ENV=production babel-node src/server.js"
-    },
-    "author": "daniel daza",
-    "license": "MIT",
-    "devDependencies": {
-        "@babel/core": "^7.7.7",
-        "@babel/node": "^7.7.7",
-        "babel-plugin-module-resolver": "^4.0.0",
-        "babel-preset-next": "^1.2.0",
-        "cross-env": "^6.0.3"
-    },
-    "dependencies": {
-        "@zeit/next-sass": "^1.0.1",
-        "axios": "^0.19.0",
-        "body-parser": "^1.19.0",
-        "cors": "^2.8.5",
-        "express": "^4.17.1",
-        "jsonwebtoken": "^8.5.1",
-        "jwt-decode": "^2.2.0",
-        "next": "^9.1.6",
-        "next-redux-wrapper": "^4.0.1",
-        "node-fetch": "^2.6.0",
-        "node-sass": "^4.13.0",
-        "react": "^16.12.0",
-        "react-dom": "^16.12.0",
-        "react-redux": "^7.1.3",
-        "redux": "^4.0.5",
-        "redux-persist": "^6.0.0",
-        "redux-thunk": "^2.3.0",
-        "socket.io-client": "^2.3.0"
-    }
-}
