--- Client/src/pages/_document.jsx
@@ -1,46 +1,0 @@
-// Dependencies
-import Document, { Head, Main, NextScript } from 'next/document';
-
-export default class MyDocument extends Document {
-    render() {
-        return (
-            <html lang="en">
-                <Head>
-                    <meta
-                        name="viewport"
-                        content="width=device-width,initial-scale=1.0,maximum-scale=1.0,minimum-scale=1.0,user-scalable=no"
-                    />
-                    <script
-                        src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-                        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-                        crossorigin="anonymous"
-                    ></script>
-                    <script
-                        src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
-                        integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
-                        crossorigin="anonymous"
-                    ></script>
-                    <link
-                        rel="stylesheet"
-                        href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-                        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
-                        crossorigin="anonymous"
-                    />
-                    <script
-                        src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
-                        integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
-                        crossorigin="anonymous"
-                    ></script>
-                    <link
-                        href="https://fonts.googleapis.com/css?family=Ma+Shan+Zheng|Oswald|Roboto+Slab&display=swap"
-                        rel="stylesheet"
-                    ></link>
-                </Head>
-                <body>
-                    <Main />
-                    <NextScript />
-                </body>
-            </html>
-        );
-    }
-}
