+++ client/src/pages/index.scss
