+++ client/src/pages/login/index.jsx
@@ -1,0 +1,109 @@
+// Dependencies
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
+
+// Actions
+import { loginUser } from '../../redux/actions/authActions';
+
+class Login extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            email: '',
+            password: '',
+            errors: {}
+        };
+        this.onChange = this.onChange.bind(this);
+        this.onSubmit = this.onSubmit.bind(this);
+    }
+
+    componentDidMount() {
+        // Check if user is already authenticated
+        if (this.props.auth.isAuthenticated) {
+            this.props.history.push('/room');
+        }
+    }
+
+    UNSAFE_componentWillReceiveProps(nextProps) {
+        // Check if user disconnect
+        if (nextProps.auth.isAuthenticated) {
+            this.props.history.push('/room');
+        }
+
+        // Check for errors
+        if (nextProps.errors) {
+            this.setState({ errors: nextProps.errors });
+        }
+    }
+
+    onChange(e) {
+        this.setState({ [e.target.name]: e.target.value });
+    }
+
+    onSubmit(e) {
+        e.preventDefault();
+
+        const User = {
+            email: this.state.email,
+            password: this.state.password
+        };
+        // Dispatch user log in
+        this.props.loginUser(User, href => {
+            this.props.history.push(href);
+        });
+    }
+
+    render() {
+        const { errors } = this.state;
+        return (
+            <div className="login">
+                <div className="container">
+                    <div className="row">
+                        <div className="col-md-8 m-auto">
+                            <h1 className="display-4 text-center">Log In</h1>
+                            <p className="lead text-center">
+                                Sign in to your Chat account
+                            </p>
+                            <form noValidate onSubmit={this.onSubmit}>
+                                <input
+                                    placeholder="Email Address"
+                                    name="email"
+                                    type="email"
+                                    value={this.state.email}
+                                    onChange={this.onChange}
+                                />
+
+                                <input
+                                    placeholder="Password"
+                                    name="password"
+                                    type="password"
+                                    value={this.state.password}
+                                    onChange={this.onChange}
+                                />
+                                <input
+                                    type="submit"
+                                    value="Submit"
+                                    className="btn btn-info btn-block mt-4"
+                                />
+                            </form>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
+}
+
+Login.propTypes = {
+    loginUser: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired
+};
+
+const mapStateToProps = state => ({
+    auth: state.auth,
+    errors: state.errors
+});
+
+export default connect(mapStateToProps, { loginUser })(withRouter(Login));
