+++ client/src/pages/register/index.jsx
@@ -1,0 +1,121 @@
+// Dependencies
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
+
+// Actions
+import { registerUser } from '../../redux/actions/authActions';
+
+class Register extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            username: '',
+            email: '',
+            password: '',
+            password2: '',
+            errors: {}
+        };
+        this.onChange = this.onChange.bind(this);
+        this.onSubmit = this.onSubmit.bind(this);
+    }
+
+    componentDidMount() {
+        // Check if user is already authenticated
+        if (this.props.auth.isAuthenticated) {
+            this.props.history.push('/room');
+        }
+    }
+
+    UNSAFE_componentWillReceiveProps(nextProps) {
+        // Check for errors
+        if (nextProps.errors) {
+            this.setState({ errors: nextProps.errors });
+        }
+    }
+
+    onChange(e) {
+        this.setState({ [e.target.name]: e.target.value });
+    }
+
+    onSubmit(e) {
+        e.preventDefault();
+
+        const newUser = {
+            username: this.state.username,
+            email: this.state.email,
+            password: this.state.password,
+            password2: this.state.password2
+        };
+        // Dispatch user register
+        this.props.registerUser(newUser, href => {
+            this.props.history.push(href);
+        });
+    }
+    render() {
+        const { errors } = this.state;
+        return (
+            <div className="register">
+                <div className="container">
+                    <div className="row">
+                        <div className="col-md-8 m-auto">
+                            <h1 className="display-4 text-center">Sign Up</h1>
+                            <p className="lead text-center">
+                                Create your DevConnector account
+                            </p>
+                            <form noValidate onSubmit={this.onSubmit}>
+                                <input
+                                    placeholder="Username"
+                                    name="username"
+                                    type="username"
+                                    value={this.state.username}
+                                    onChange={this.onChange}
+                                />
+                                <input
+                                    placeholder="Email Address"
+                                    name="email"
+                                    type="email"
+                                    value={this.state.email}
+                                    onChange={this.onChange}
+                                />
+                                <input
+                                    placeholder="Password"
+                                    name="password"
+                                    type="password"
+                                    value={this.state.password}
+                                    onChange={this.onChange}
+                                />
+                                <input
+                                    placeholder="Password"
+                                    name="password2"
+                                    type="password"
+                                    value={this.state.password2}
+                                    onChange={this.onChange}
+                                    error={errors.password2}
+                                />
+                                <input
+                                    type="submit"
+                                    value="Submit"
+                                    className="btn btn-info btn-block mt-4"
+                                />
+                            </form>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
+}
+
+Register.propTypes = {
+    registerUser: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired
+};
+
+const mapStateToProps = state => ({
+    auth: state.auth,
+    errors: state.errors
+});
+
+export default connect(mapStateToProps, { registerUser })(withRouter(Register));
