+++ client/src/pages/room/index.scss
